move ProductCard into /components/Common
